transferring logic to a separate component
